Reduce duplicate code among tool scripts by using util.js.
