+++ tools/cross-shard/sendCheckinTx.js
@@ -1,11 +1,5 @@
 const path = require('path');
-const _ = require('lodash');
-const axios = require('axios');
-const {sleep} = require('sleep');
-const ainUtil = require('@ainblockchain/ain-util');
-const ChainUtil = require('../../common/chain-util');
-
-const CURRENT_PROTOCOL_VERSION = require('../../package.json').version;
+const { signTx, signAndSendTx, confirmTransaction } = require('../util');
 let config = {};
 
 function buildPayloadTxBody(fromAddr, toAddr, tokenAmount, timestamp) {
@@ -36,61 +30,6 @@ function buildTriggerTxBody(address, payload, timestamp) {
   }
 }
 
-function signTx(txBody, privateKey) {
-  const keyBuffer = Buffer.from(privateKey, 'hex');
-  const sig = ainUtil.ecSignTransaction(txBody, keyBuffer);
-  const sigBuffer = ainUtil.toBuffer(sig);
-  const lenHash = sigBuffer.length - 65;
-  const hashedData = sigBuffer.slice(0, lenHash);
-  const txHash = '0x' + hashedData.toString('hex');
-  return {
-    txHash,
-    signedTx: {
-      tx_body: txBody,
-      signature: sig,
-      protoVer: CURRENT_PROTOCOL_VERSION,
-    }
-  };
-}
-
-function signAndSendTx(endpointUrl, txBody, privateKey) {
-  console.log('\n*** signAndSendTx():');
-  const {txHash, signedTx} = signTx(txBody, privateKey);
-  return axios.post(
-    `${endpointUrl}/json-rpc`,
-    {
-      method: 'ain_sendSignedTransaction',
-      params: signedTx,
-      jsonrpc: '2.0',
-      id: 0
-    }
-  ).then((resp) => {
-    const success = !ChainUtil.transactionFailed(_.get(resp, 'data.result'), null);
-    console.log(`result: ${JSON.stringify(success, null, 2)}`);
-    return {txHash, signedTx, success};
-  }).catch((err) => {
-    console.log(`Failed to send transaction: ${err}`);
-    return {errMsg: err.message};
-  });
-}
-
-async function sendGetTxByHashRequest(endpointUrl, txHash) {
-  return await axios.post(
-    `${endpointUrl}/json-rpc`,
-    {
-      method: 'ain_getTransactionByHash',
-      params: {
-        protoVer: CURRENT_PROTOCOL_VERSION,
-        hash: txHash,
-      },
-      jsonrpc: '2.0',
-      id: 0
-    }
-  ).then(function(resp) {
-    return _.get(resp, 'data.result.result', null);
-  });
-}
-
 async function sendTransaction() {
   console.log('\n*** sendTransaction():');
   const timestamp = Date.now();
@@ -110,29 +49,12 @@ async function sendTransaction() {
   return {timestamp, txInfo};
 }
 
-async function confirmTransaction(timestamp, txHash) {
-  console.log('\n*** confirmTransaction():');
-  console.log(`txHash: ${txHash}`);
-  let iteration = 0;
-  let result = null;
-  while (true) {
-    iteration++;
-    result = await sendGetTxByHashRequest(config.endpointUrl, txHash);
-    sleep(1);
-    if (_.get(result, 'is_finalized')) {
-      break;
-    }
-  }
-  console.log(`iteration = ${iteration}, result: ${JSON.stringify(result, null, 2)}`);
-  console.log(`elapsed time (ms) = ${result.finalized_at - timestamp}`);
-}
-
 async function sendCheckinTransaction() {
   console.log('\n*** sendTransaction():');
   console.log(`config: ${JSON.stringify(config, null, 2)}`);
   const {timestamp, txInfo} = await sendTransaction();
   if (txInfo.success) {
-    await confirmTransaction(timestamp, txInfo.txHash);
+    await confirmTransaction(config.endpointUrl, timestamp, txInfo.txHash);
   }
 }
